\name{SpatialAlign-package}
\alias{SpatialAlign-package}
\alias{SpatialAlign}
\docType{package}
\title{
\packageTitle{SpatialAlign}
}
\description{
\packageDescription{SpatialAlign}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpatialAlign}
\packageIndices{SpatialAlign}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SpatialAlign}

Maintainer: \packageMaintainer{SpatialAlign}
}
